pyproject.toml
[project]
name = "jdocmunch-mcp"
version = "1.1.0"
description = "Token-efficient MCP server for structured documentation retrieval via section-level indexing"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.10.0,<2.0.0",
    "httpx>=0.27.0",
    "pathspec>=0.12.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40.0"]
gemini = ["google-generativeai>=0.8.0"]

[project.scripts]
jdocmunch-mcp = "jdocmunch_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jdocmunch_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
